+++ Controller/Content/View.php
@@ -41,6 +41,9 @@ class View extends AbstractAction
         } finally {
             if (!empty($e)) {
                 $this->logger->error($e);
+                if ($this->dataHelper->isDebugMode()) {
+                    throw $e;
+                }
             }
         }