+++ Model/ItemProvider/PageBuilderSitemap.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 namespace Goomento\PageBuilder\Model\ItemProvider;
 
 use Goomento\PageBuilder\Api\Data\ContentInterface;
+use Goomento\PageBuilder\Helper\Data;
 use Magento\Sitemap\Model\ItemProvider\CmsPageConfigReader;
 use Magento\Sitemap\Model\ItemProvider\ItemProviderInterface;
 use Magento\Sitemap\Model\SitemapItemInterfaceFactory;
@@ -34,23 +35,30 @@ class PageBuilderSitemap implements ItemProviderInterface
      * @var Logger
      */
     private $logger;
+    /**
+     * @var Data
+     */
+    private $dataHelper;
 
     /**
      * @param CollectionFactory $collectionFactory
      * @param SitemapItemInterfaceFactory $itemFactory
      * @param CmsPageConfigReader $configReader
+     * @param Data $dataHelper
      * @param Logger $logger
      */
     public function __construct(
         CollectionFactory $collectionFactory,
         SitemapItemInterfaceFactory $itemFactory,
         CmsPageConfigReader $configReader,
+        Data $dataHelper,
         Logger $logger
     )
     {
         $this->collectionFactory = $collectionFactory;
         $this->itemFactory = $itemFactory;
         $this->configReader = $configReader;
+        $this->dataHelper = $dataHelper;
         $this->logger = $logger;
     }
 
@@ -60,10 +68,15 @@ class PageBuilderSitemap implements ItemProviderInterface
     public function getItems($storeId)
     {
         try {
+            // Avoid to indexing the pages if disabled
+            if (!$this->dataHelper->isActive()) {
+                return [];
+            }
+
             /** @var Collection $collection */
             $collection = $this->collectionFactory->create();
             $collection
-                ->addStoreFilter((int) $storeId, false)
+                ->addStoreFilter((int) $storeId, true)
                 ->addFilter(ContentInterface::TYPE, ContentInterface::TYPE_PAGE)
                 ->addFilter(ContentInterface::STATUS, ContentInterface::STATUS_PUBLISHED)
                 ->addFilter(ContentInterface::ENABLED, ContentInterface::ENABLED);