+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Magento 2 Goomento Page Builder
 
+## 0.2.6: Added and updated widgets
+- Fixed minor bugs
+- Added debug mode configuration
+- Updated Banner, Banner Slider, Product + Product Slider widgets
+- Added Facebook widgets, Magento widgets
+
 ## 0.2.5: Updated document
 - Fixed minor bugs
 - Updated document